refactor: extracted the reusable components
