+++ wondrous-app/components/OnboardingDao/OnboardingForm/index.tsx
@@ -1,0 +1,1 @@
+export { default } from './component';
